requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
ccxt==4.4.72
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
fonttools==4.56.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
multidict==6.3.1
narwhals==1.33.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
propcache==0.3.1
protobuf==5.29.4
pyarrow==19.0.1
pycares==4.5.0
pycparser==2.22
pydeck==0.9.1
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
smmap==5.0.2
streamlit==1.44.1
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
watchdog==6.0.0
yarl==1.18.3
